fix: refactor to class based decorators
